trying to fix components
