+++ src/tempFOLDER/Footer/Footer.css
@@ -1,0 +1,80 @@
+footer {
+    background-color: #111111;
+    margin-top: 5rem;
+    color: white;
+    font-family: 'Poppins', sans-serif;
+}
+
+.footer_container {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    padding: 5rem 7.5rem;
+    align-content: center;
+}
+
+.footer_container .social {
+    display: flex;
+    flex-direction: column;
+}
+
+.footer_container .social .links {
+    display: flex;
+    flex-direction: row;
+}
+
+.footer_container .social .links a {
+    padding: 6px 12px;
+    margin-right: 10px;
+    border-radius: 50%;
+    background-color: white;
+}
+
+.footer_container .content-list {
+    display: grid;
+    grid-template-columns: auto auto auto;
+
+
+}
+
+.footer_container .content-list .item {
+    display: flex;
+    flex-direction: column;
+    margin-right: 10rem;
+}
+
+.footer_container .content-list .item ul{
+    list-style: none;
+}
+
+.footer_container .content-list .item li {
+    margin: 0;
+    font-weight: 500;
+    font-size: 14px;
+}
+
+.footer_container h5 {
+    font-weight: 500;
+    text-transform: uppercase;
+    font-size: 12px;
+    margin-bottom: 20px;
+}
+
+@media only screen and (max-width: 768px) {
+    .footer_container{
+        flex-direction: column;
+        padding: 5rem 2rem ;
+    }
+
+    .footer_container .content-list {
+        grid-template-columns: auto auto;
+        gap: 2rem;
+    }
+    .footer_container h5{
+        margin-bottom: 5px;
+    }
+
+    .footer_container .social{
+        margin-top: 2rem;
+    }
+}
