+++ src/tempFOLDER/ImageSlider/images.js
@@ -1,0 +1,35 @@
+import img1 from '../../assets/img1.jpg'
+import img2 from '../../assets/img2.jpg'
+import img3 from '../../assets/img3.jpg'
+import img4 from '../../assets/img4.jpg'
+import img5 from '../../assets/img5.jpg'
+import img6 from '../../assets/img6.jpg'
+
+const images = [
+    {
+       name: "devansh",
+       img: img1, 
+    },
+    {
+        name: "devansh",
+        img: img2, 
+     },
+     {
+        name: "devansh",
+        img: img3, 
+     },
+     {
+        name: "devansh",
+        img: img4, 
+     },
+     {
+        name: "devansh",
+        img: img5, 
+     },
+     {
+        name: "devansh",
+        img: img6, 
+     },
+]
+
+export default images;
