+++ src/tempFOLDER/VotingPage/DesignModal.jsx
