+++ src/tempFOLDER/VotingPage/Dropzone.css
