+++ src/tempFOLDER/VotingPage/Dropzone.jsx
